requirements.txt
absl-py==2.3.1
-e git+https://github.com/Gab-Angel/SDR_Agent@3875ab41787c9ebc47f44c0973f63482de375bd5#egg=agente_SDR
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.22.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.0
bidict==0.23.1
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.2
cryptography==45.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
dell-recovery==0.0.0
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
et_xmlfile==2.0.0
evolutionapi==0.1.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fasteners==0.14.1
fastjsonschema==2.21.1
filetype==1.2.0
Flask==3.1.1
flask-ngrok==0.0.25
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.6.2
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.9.0
google-api-core==2.25.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.2
greenlet==3.2.3
gritql==0.2.0
groq==0.26.0
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==23.5.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-cli==0.0.36
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langserve==0.3.1
langsmith==0.3.45
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.4.4
narwhals==1.41.1
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.86.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.10.18
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==24.2
PAM==0.4.2
pandas==2.3.0
paramiko==2.9.3
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.8.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-mock==3.14.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
realtime==2.5.3
redis==6.2.0
referencing==0.36.2
reportlab==3.6.8
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
SecretStorage==3.3.1
selenium==4.33.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.46.2
storage3==0.12.0
streamlit==1.45.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
systemd-python==234
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.4.0
usb-creator==0.3.7
uvicorn==0.34.3
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.20.0
zipp==1.0.0
zstandard==0.23.0
